Set MessageReader member to correct type
MessageReader member pos_ was set to int when it should have been
size_t.

PiperOrigin-RevId: 267397467

+++ asylo/platform/primitives/util/message.h
@@ -247,7 +247,7 @@ class MessageReader {
 
  private:
   std::vector<std::pair<std::unique_ptr<char[]>, size_t>> extents_;
-  int pos_ = 0;
+  size_t pos_ = 0;
 };
 
 }  // namespace primitives
